build(web): add a dummy Module and pass it version and patchLevel
update.js (simulator.html) expects to find Module and these three methods when
it loads. However since we now load the module asynchronously they may not exist
yet so we provide it this dummy Module object instead.

An alternative was to load the script later in ModuleAsync.onRuntimeInitialized
but the CORS on the website do not allow this.